Cleanup of the use of properties; prefer int to unsigned int in properties.
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/RobWork/trunk@349 494c09c7-a67d-49d4-9938-151fb1965be8

+++ src/rw/common/StringUtil.cpp
@@ -76,7 +76,7 @@ std::string StringUtil::GetFileExtension(
 std::string StringUtil::ToUpper(const std::string& str)
 {
     std::string result(str);
-	for(unsigned int i=0; i < str.length(); i++)
+	for(size_t i=0; i < str.length(); i++)
 		result[i] = toupper(str[i]);
 
     return result;
@@ -85,7 +85,7 @@ std::string StringUtil::ToUpper(const std::string& str)
 std::string StringUtil::ToLower(const std::string& str)
 {
     std::string result(str);
-	for(unsigned int i=0; i < str.length(); i++)
+	for(size_t i=0; i < str.length(); i++)
 		result[i] = tolower(str[i]);
 
     return result;
